+++ example/lib/widgets/wrap.dart
@@ -30,6 +30,7 @@ class AzWrap extends StatelessWidget {
         Key? key,
       }) : super(key: key);
 
+
   AzWrap axis(Axis direction){
     _direction = direction;
     return this;