+++ lib/widgets/row.dart
@@ -21,6 +21,20 @@ class AzRow extends StatelessWidget{
   TextBaseline? _textBaseline;
 
   AzRow(this.children, {Key? key}) : super(key: key);
+  AzRow internalUseOfAz({
+    thisMainAxisAlignment,
+    thisCrossAxisAlignment,
+    thisTextDirection,
+    thisVerticalDirection,
+    thisTextBaseline,
+  }){
+    _mainAxisAlignment = thisMainAxisAlignment;
+    _crossAxisAlignment = thisCrossAxisAlignment;
+    _textDirection = thisTextDirection;
+    _verticalDirection = thisVerticalDirection;
+    _textBaseline = thisTextBaseline;
+    return this;
+  }
 
   AzRow crossStart(){
     _crossAxisAlignment = CrossAxisAlignment.start;