+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-## 0.0.5/0.0.6
+## 0.0.11
+* bugs fix
+## 0.0.9
+* azCard() extension now can use card properties if defined
+* azCenter() extension now can use center properties if defined
+* azColumn() extension now can use column properties if defined
+* azIcon() extension now can use common Icon properties if defined
+* azInkWell() extension now can use two common InkWell properties if defined
+* azRow() extension now can use row properties if defined
+* azText() extension now can use some common text properties if defined
+* Do not use AzImage() it is not rebuild so showing dummy images
+## 0.0.6
 * Some InkWell Properties added
 * Methods changes container to toContainer, center to toCenter to make suggestions easy
 * to use hex code a method is added on string fromHex
@@ -15,3 +26,4 @@
 * Try to fixing analyses issues.
 ## 0.0.1
 * A package for easy & smooth development of UI.
+* This package covers the necessary properties.